feat(statement): split phase and move create statement data to separate file
feat(createStatementData): replace object with map

feat(createStatementData): move to separate file
